+++ vitest.config.ts
@@ -1,15 +1,13 @@
-import { defineWorkersConfig } from "@cloudflare/vitest-pool-workers/config";
+import { cloudflareTest } from "@cloudflare/vitest-pool-workers";
+import { defineConfig } from "vitest/config";
 
-export default defineWorkersConfig({
+export default defineConfig({
+	plugins: [
+		cloudflareTest({
+			wrangler: { configPath: "./wrangler.jsonc" },
+		}),
+	],
 	test: {
 		globals: true,
-		poolOptions: {
-			workers: {
-				main: "./src/test-worker.ts",
-				miniflare: {
-					compatibilityDate: "2024-12-25",
-				},
-			},
-		},
 	},
 });